+++ .vscode/settings.json
@@ -5,6 +5,7 @@
     "agentic",
     "Andreessen",
     "apidevtools",
+    "Bleu",
     "CBRNE",
     "chatbots",
     "chatcompletion",
@@ -25,6 +26,7 @@
     "logprobs",
     "Mateo",
     "mitigations",
+    "ngrams",
     "openai",
     "overreliance",
     "OWASP",
